backend/requirements.txt
aiofiles
aiohttp
aiortc
albumentations
altair
amqp
audio-utils
av
azure-core
azure-cosmos
azure-identity
azure-storage-blob
azure-storage-queue
beautifulsoup4
celery
dlib
dotenv
easydict
eyeD3
face-recognition
face_recognition_models
faiss-cpu
fastapi
ffmpy
filetype
GitPython
google-api-python-client
google-auth
google-cloud-texttospeech
google-generativeai
google-genai
grpcio
gTTS
huggingface-hub
imageio
insightface
ipykernel
ipython
jedi
Jinja2
joblib
jsonschema
jupyter_client
jupyter_core
lxml
matplotlib
ml_dtypes
msal
msal-extensions
mutagen
mysql-connector-python
networkx
numba
numpy
onnx
onnxruntime
openai
openai-whisper
opencv-python
opencv-python-headless
pandas
pdfminer.six
pdfplumber
pillow
pydantic
pydeck
pydub
pygame
PyJWT
pylibsrtp
pymongo
pyOpenSSL
pypdfium2
python-docx
python-dotenv
python-multipart
redis
requests
rich
rq
safetensors
scikit-image
scikit-learn
scipy
sentence-transformers
sounddevice
soundfile
SpeechRecognition
starlette
streamlit
streamlit-webrtc
sympy
tenacity
textual
tiktoken
tokenizers
torch
tqdm
transformers

uvicorn
watchdog
webrtcvad
websockets
